+++ src/controllers/testimonial-controller.js
@@ -1,4 +1,5 @@
 const testimonialService = require('../services/testimonial-service')
+const { uploadFile } = require('../services/s3-service');
 
 
 const getAllTestimonials = async (req, res, next) => {
@@ -18,6 +19,31 @@ const getAllTestimonials = async (req, res, next) => {
 
 }
 
+const createNewTestimony = async(req, res) => {
+    // Uploads image to AWS S3 service and extract de key value
+    const imageFileCheck = async function () {
+        if(req.file){
+            const { key } = await uploadFile( req.file );
+            return key
+        } else {
+            return ""
+        }
+    }
+
+    const image = await imageFileCheck()
+  
+    const testimonySaved = await testimonialService.createTestimony({
+      ...req.body,
+      image
+    })
+  
+    res.status(201).json({
+      ok: true,
+      entry: testimonySaved
+    })
+  }
+
 module.exports = {
-    getAllTestimonials
+    getAllTestimonials,
+    createNewTestimony
 }
