--- src/repositories/testimony-repository.js
@@ -1,10 +1,0 @@
-const { Testimony, sequelize } = require('../models')
-
-const createTestimony = async (testimony) => {
-    const testimonyStored = await Testimony.create(testimony)
-    return testimonyStored
-}
-
-module.exports = {
-    createTestimony
-}
