+++ src/services/testimonial-service.js
@@ -12,7 +12,13 @@ const getAllTestimonials = async (limit,page) => {
     return testimonialsForPage
 }
 
+const createTestimony = async( testimony ) =>{
+    const testimonyStored = await testimonialRepository.createTestimony(testimony);
+    return testimonyStored;
+  }
+
 
 module.exports= {
     getAllTestimonials,
+    createTestimony
 }
