--- src/services/testimony-service.js
@@ -1,10 +1,0 @@
-const testimonyRepository = require('../repositories/testimony-repository')
-
-const createTestimony = async( testimony ) =>{
-    const testimonyStored = await testimonyRepository.createTestimony(testimony);
-    return testimonyStored;
-  }
-  
-  module.exports ={
-    createTestimony
-  }
